requirements.txt
playwright
requests
python-dotenv